Config files are now loaded via enviroment variables.
Starting the server with no config arguments causes the application to look for an `application.local.conf` file in the current working directory, followed any application files in the resources folder. In addition, it adds any command line and environment variables to the config, with a higher priority than the config files.

The legacy behavior is still preserved when manually specifying a config file, no auto-loading or config merging is performed.